[quantization] Split config classes
This commit splits config classes.

TICO-DCO-1.0-Signed-off-by: seongwoo <mhs4670go@naver.com>
