+++ test/pt2_to_qcircle_test/builder.py
@@ -18,7 +18,7 @@ from pathlib import Path
 
 import tico
 from tico.experimental.quantization import convert, prepare
-from tico.experimental.quantization.config import PT2EConfig
+from tico.experimental.quantization.config.pt2e import PT2EConfig
 from tico.experimental.quantization.evaluation.backend import BACKEND
 from tico.experimental.quantization.evaluation.evaluate import evaluate
