+++ test/quantization/algorithm/test_gptq.py
@@ -19,7 +19,8 @@ import tico
 import torch
 
 from tico.experimental.quantization import convert, prepare
-from tico.experimental.quantization.config import GPTQConfig, PT2EConfig
+from tico.experimental.quantization.config.gptq import GPTQConfig
+from tico.experimental.quantization.config.pt2e import PT2EConfig
 from tico.experimental.quantization.evaluation.evaluate import BACKEND, evaluate
 
 IS_INTERNAL_TEST = os.environ.get("RUN_INTERNAL_TESTS", "0") == "1"
